compress: one temp_store, one make_wave, one boot_server
chat.rs and radio.rs each carried an identical temp_store and a make_wave
differing only in an unread goal field; chat.rs and memory.rs each booted
the same server, differing only in whether they kept the inbox receiver.
A cfg(test) fixtures module under lf/commands holds all three — they are
made of loopflow's own types, so loopflow-test-support cannot host them
without depending back on this crate.

Also note in scratch/questions.md why item_line is left duplicated.
