+++ rust/loopflow/src/lf/commands/chat.rs
@@ -319,44 +319,15 @@ mod tests {
 
     use time::OffsetDateTime;
 
+    use crate::lf::commands::fixtures::{boot_server, make_wave, temp_store};
     use crate::lfd::id::LfdId;
     use crate::lfd::types::{
-        Session, SessionStatus, SessionUse, WaveStatus, WAVE_SERVER_ENDPOINT_ENV,
-        WAVE_SERVER_PID_ENV, WAVE_SERVER_SOURCE,
+        Session, SessionStatus, SessionUse, WAVE_SERVER_ENDPOINT_ENV, WAVE_SERVER_PID_ENV,
+        WAVE_SERVER_SOURCE,
     };
-    use crate::lfdb::{open_store, StorageConfig};
     use crate::wave::journal::{EventKind, MessageOp};
-    use crate::wave::runtime::{InboxItem, WaveRuntime};
+    use crate::wave::runtime::InboxItem;
     use crate::wave::server;
-
-    async fn temp_store(dir: &Path) -> SharedStore {
-        Arc::new(
-            open_store(&StorageConfig::sqlite(dir.join("lfd.db")))
-                .await
-                .expect("open sqlite store"),
-        )
-    }
-
-    fn make_wave(name: &str, repo: &Path, parent: Option<&LfdId>) -> Wave {
-        Wave {
-            id: LfdId::new(),
-            name: name.to_string(),
-            goal: "ship-roadmap".to_string(),
-            metrics: Vec::new(),
-            repo: repo.display().to_string(),
-            worktree: String::new(),
-            branch: String::new(),
-            status: WaveStatus::Idle,
-            iteration: 0,
-            cycle_start_iteration: 0,
-            direction: Vec::new(),
-            area: Vec::new(),
-            paused: false,
-            created_at: Some(OffsetDateTime::now_utc()),
-            workers: 1,
-            parent_wave_id: parent.cloned(),
-        }
-    }
 
     /// A live wave_server WaveAgent row carrying `endpoint` in its env — the
     /// shape `lf loop` registers (see crate::wave::registry).
@@ -388,33 +359,6 @@ mod tests {
             created_at: now,
             completion_token: None,
         }
-    }
-
-    /// Boot the HTTP surface over a runtime (the wave/mod.rs harness pattern).
-    async fn boot_server(
-        origin: &Path,
-        wave: &str,
-    ) -> (
-        String,
-        Arc<WaveRuntime>,
-        tokio::sync::broadcast::Receiver<InboxItem>,
-    ) {
-        let runtime =
-            WaveRuntime::open(wave.to_string(), origin.to_path_buf()).expect("open runtime");
-        let inbox_rx = runtime.subscribe_inbox();
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let app = server::router(
-            runtime.clone(),
-            server::ResidentDoor::new("test-token"),
-            server::SubagentDoor::new(),
-            None,
-            None,
-        );
-        tokio::spawn(async move {
-            axum::serve(listener, app).await.ok();
-        });
-        (addr.to_string(), runtime, inbox_rx)
     }
 
     /// Env-context targeting: LFD_WAVE_ID names the wave; the endpoint comes
